pyproject.toml
[tool.poetry]
name = "wallet-action"
version = "0.1.0"
description = ""
authors = ["Kem <diyeqo@icloud.com>"]
readme = "README.md"
packages = [{include = "wallet_action"}]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.2.0"
aiohttp = "^3.8.4"
aiogram = {extras = ["test"], version = "^2.25.1"}
aiofiles = "^23.1.0"
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"
pytest-asyncio = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
